Reworking formats to add animations
Note: not finished or tested.
I finally got animations (api wise) the way I want them; they use
genenrics and some fancy stuff to make them work, and they work without
ties to sending anything, just ties to the sponge scheduler. Should be
good. Also made a quick internal change to allow variable parsing to be
a bit neater with formats, so as to not have to pass maps. Should be
good. I will probably remove the valueholder class if I can find an
equivilant in the api. Need to do some hardcore testing with this
though, and I should setup a test mechanism for this soon.